made the filter page into a reusable widget.
